backend/requirements_dev.txt
-r requirements.txt

coverage
pylint
pylint-django
pytest
pytest-django
